# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModel:
    properties:
      bodyType:
        type: string
      brand:
        type: string
      class:
        type: string
      createdAt:
        type: string
      engineVolume:
        type: number
      features:
        items:
          type: string
        type: array
      fuelType:
        type: string
      id:
        type: string
      imageStorageUrls:
        items:
          type: string
        type: array
      model:
        type: string
      rangeKm:
        type: integer
      seats:
        type: integer
      transmission:
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelCreateRequest:
    properties:
      bodyType:
        type: string
      brand:
        type: string
      class:
        type: string
      engineVolume:
        type: number
      features:
        items:
          type: string
        type: array
      fuelType:
        type: string
      imageStorageKeys:
        items:
          type: string
        type: array
      model:
        type: string
      rangeKm:
        type: integer
      seats:
        type: integer
      transmission:
        type: string
      year:
        type: integer
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelGetResponse:
    properties:
      carModel:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModel'
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelUpdateRequest:
    properties:
      bodyType:
        type: string
      brand:
        type: string
      class:
        type: string
      engineVolume:
        type: number
      features:
        items:
          type: string
        type: array
      fuelType:
        type: string
      imageStorageKeys:
        items:
          type: string
        type: array
      model:
        type: string
      rangeKm:
        type: integer
      seats:
        type: integer
      transmission:
        type: string
      year:
        type: integer
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelsResponse:
    properties:
      carModels:
        items:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModel'
        type: array
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail'
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse:
    properties:
      id:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadData:
    properties:
      objectKey:
        type: string
      presignedPutURL:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadResponse:
    properties:
      uploadData:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadData'
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@carrental.kz
    name: API Support
  description: API Gateway for the Kazakhstan carsharing platform. Handles auth, users,
    fleet, zones, insurance, and maintenance.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Car Rental API Gateway
  version: "1.0"
paths:
...
  /car-models:
    get:
      description: Returns a filtered, paginated list of car models.
      parameters:
      - description: Brand filter (e.g. Hyundai)
        in: query
        name: brand
        type: string
      - description: Model name filter
        in: query
        name: model
        type: string
      - description: Fuel type (petrol, diesel, electric, hybrid)
        in: query
        name: fuelType
        type: string
      - description: Transmission (automatic, manual)
        in: query
        name: transmission
        type: string
      - description: Body type (sedan, suv, hatchback…)
        in: query
        name: bodyType
        type: string
      - description: Class (economy, comfort, business)
        in: query
        name: class
        type: string
      - description: Minimum seat count
        in: query
        name: minSeats
        type: integer
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List car models
      tags:
      - car-models
    post:
      consumes:
      - application/json
      description: Registers a new car model (e.g. Hyundai Solaris 2022). Returns
        the new model's UUID.
      parameters:
      - description: Car model payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create car model
      tags:
      - car-models
  /car-models/{id}:
    delete:
      parameters:
      - description: Car model UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete car model
      tags:
      - car-models
    get:
      parameters:
      - description: Car model UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelGetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get car model by ID
      tags:
      - car-models
    patch:
      consumes:
      - application/json
      parameters:
      - description: Car model UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CarModelUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update car model
      tags:
      - car-models
  /car-models/image-upload:
    get:
      description: Returns a pre-signed S3-compatible URL and object key for uploading
        a car model image.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ImageUploadResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pre-signed image upload URL
      tags:
      - car-models
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"